Allow field renames (#49)
* Allow fields to be renamed
Valid JSON property names are not always valid Go identifiers. A common example is a property name that starts with a number. This cannot be turned into a field in the Go struct generated by this tool.
With this change you can supply key value pairs to rename JSON property names to something that is a valid Go identifier. For example if you had a property name in your JSONschema file called "16_9" you could call go generate like this:
```
//go:generate jstransform -renameFields 16_9=SixteenByNine /path/to/the/schema.json
```